feat: Update drizzle configuration to use new database URL
The `drizzle.config.js` file was modified to update the database URL used in the application. The `url` property in the `dbCredentials` object was changed from `process.env.NEXT_PUBLIC_DATABASE_URL` to `process.env.NEXT_PUBLIC_DRIZZLE_DATABASE_URL`. This change ensures that the application connects to the correct database.

Please review and merge this change.